spark/requirements_spark.txt
pyspark==3.5.1
kafka-python==2.0.2
psycopg2-binary==2.9.9